fix bracket height with match overflow
Because final winner container and bracket title were absolutely
positioned relative to the final match box, a title longer than one line
would overflow the bracket bounds. Instead, fix the height of the
bracket to 100px and let the matches themselves overflow. This way the
winner container and bracket logo can remain in the flow and position
the bracket accordingly