Get num plays through leaderboard service
Use the leaderboard service to get num plays for a given bracket, and
add a num_plays attribute to Bracket. When using the leaderboard service
to get num plays in BracketRepo, a segmentation fault occurs because
leaderboard service uses bracket and play repos internally, leading to a
circular dependency. Need to fix somehow