paid tournament entry (#343)
* Clear tournaments entries for paid brackets

This commit updates TournamentEntryService to always clear
tournament entries before adding a new one, even if the bracket is paid.
This enforces a single-play policiy for paid tournaments. The next
series of commits will allow users to submit subsequent plays for free
after paying for the first one.

* disable BracketProductHooks fee hooks

* create new function to check if play should be marked paid

* mark play as paid if user has existing paid play

* object-based permission check for wpbb_play_paid_bracket_for_free perm

* unified permission to check if fee is required

Rename wpbb_play_paid_bracket_for_free permission to
wpbb_play_bracket_for_free and have it return true if the bracket is
unpaid. This simplifies the logic surrounding whether to require payment
or not.

* add comment